Generalize the loader system & reorganize the architecture
