+++ lib/Diaporama.js
@@ -3,7 +3,9 @@ var EventEmitter = require("events").EventEmitter;
 var assign = require("object-assign");
 var BezierEasing = require("bezier-easing");
 
-var hash = require("./loaders/hash");
+var loader = require("./loader");
+var MediaLoader = require("./MediaLoader");
+var hash = require("./hashResource");
 var Channels = require("./Channels");
 var findTransitionByName = require("./findTransitionByName");
 var DiaporamaRenderingCanvas = require("./DiaporamaRenderingCanvas");
@@ -41,12 +43,12 @@ function guessMode (mode) {
   return mode;
 }
 
-function createRendering (container, background, mode) {
+function createRendering (media, container, background, mode) {
   switch (mode) {
   case "webgl":
-    return new DiaporamaRenderingCanvas(container, background);
+    return new DiaporamaRenderingCanvas(media, container, background);
   case "dom":
-    return new DiaporamaRenderingDOM(container, background);
+    return new DiaporamaRenderingDOM(media, container, background);
   }
   return null;
 }
@@ -81,7 +83,8 @@ function Diaporama (container, data, opts) {
   if (!this._renderingMode) {
     throw new Error("No rendering is suitable for renderingMode="+opts.renderingMode);
   }
-  this._rendering = createRendering(container, opts.backgroundColor || this._backgroundColor, this._renderingMode);
+  var media = new MediaLoader(loader.fallbackBlack(loader.timeout(loader.loader, opts.networkTimeout || 60000)));
+  this._rendering = createRendering(media, container, opts.backgroundColor || this._backgroundColor, this._renderingMode);
 
   // Delete handled options that are not possible to set.
   delete opts.renderingMode;
@@ -296,9 +299,6 @@ Diaporama.prototype = assign({}, EventEmitter.prototype, {
     }
 
     function watch () {
-      Channels.SLIDE2D.forEach(function (channel) {
-        rendering.getChannelContext(channel).setImgsCache(media.resources);
-      });
 
       if (ready.length === Object.keys(toload).length) {
         canplay();
@@ -354,7 +354,7 @@ Diaporama.prototype = assign({}, EventEmitter.prototype, {
   _loadTransitions: function () {
     var store = this._rendering.transitions;
     if (!store) return;
-    var transitionsDataset = this._data.transitions || this.GlslTransitions || {}; // FIXME: GlslTransitions deprecated
+    var transitionsDataset = this._data.transitions || {};
     this._data.timeline.forEach(function (item) {
       var name = item.transitionNext && item.transitionNext.name;
       if (name && !store.has(name)) {
