+++ lib/DiaporamaRenderingCanvas/index.js
@@ -7,10 +7,10 @@ var SegmentTransition = require("./SegmentGlslTransition");
 var SegmentKenBurns = require("./SegmentKenBurnsWebGL");
 var SegmentSlide2d = require("../SegmentSlide2d");
 var StoreTransitions = require("./StoreTransitions");
-var MediaLoader = require("../MediaLoader");
 var Channels = require("../Channels");
 
-function DiaporamaRenderingCanvas (container, bgColor) {
+function DiaporamaRenderingCanvas (media, container, bgColor) {
+  this.media = media;
   this._container = container;
   this.bg = bgColor.concat([ 1 ]);
   this._canvas = null;
@@ -20,6 +20,8 @@ function DiaporamaRenderingCanvas (container, bgColor) {
   this._fbos = {};
   this._currentChannel = null;
   var shape = this.shape = [ 1, 1 ];
+
+  var resolveImage = media.getImageResolver();
 
   var canvas = document.createElement("canvas");
   var gl = canvas.getContext("webgl");
@@ -44,11 +46,10 @@ function DiaporamaRenderingCanvas (container, bgColor) {
     var texture = createTexture(gl, canvas);
     texture.minFilter = texture.magFilter = gl.LINEAR;
     this._textures[c] = texture;
-    this._attachContext(c, Slide2d(ctx));
+    this._attachContext(c, Slide2d(ctx, resolveImage));
   }, this);
 
   this._container.appendChild(this._canvas = canvas);
-  this.media = new MediaLoader();
   this.transitions = new StoreTransitions(this.getChannelContext(Channels.TRANSITION));
 }
