+++ lib/DiaporamaRenderingDOM/index.js
@@ -2,12 +2,12 @@ var KenBurnsDOM = require("kenburns-dom");
 var Slide2d = require("slide2d");
 
 var Channels = require("../Channels");
-var MediaLoader = require("../MediaLoader");
 var SegmentTransition = require("./SegmentDomTransition");
 var SegmentKenBurns = require("./SegmentKenBurnsDOM");
 var SegmentSlide2d = require("../SegmentSlide2d");
 
-function DiaporamaRenderingDOM (container, bgColor) {
+function DiaporamaRenderingDOM (media, container, bgColor) {
+  this.media = media;
   this._container = container;
   var curPos = this._container.style.position;
   if (!curPos || curPos==="static")
@@ -16,11 +16,10 @@ function DiaporamaRenderingDOM (container, bgColor) {
   this._nodes = {};
   this._ctxs = {};
   this.createDOM();
-  this.media = new MediaLoader();
+  this.resolveImage = media.getImageResolver();
 }
 
 DiaporamaRenderingDOM.prototype = {
-  // FIXME: implement them
   SegmentTransition: SegmentTransition,
   SegmentKenBurns: SegmentKenBurns,
   SegmentSlide2d: SegmentSlide2d,
@@ -96,6 +95,7 @@ DiaporamaRenderingDOM.prototype = {
   },
 
   createDOM: function () {
+    var resolveImage = this.resolveImage;
     var div = document.createElement("div");
     this._attachChannel(Channels.TRANSITION, div, div);
 
@@ -109,7 +109,7 @@ DiaporamaRenderingDOM.prototype = {
     Channels.SLIDE2D.forEach(function (c) {
       var canvas = document.createElement("canvas");
       var ctx = canvas.getContext("2d");
-      this._attachChannel(c, canvas, Slide2d(ctx));
+      this._attachChannel(c, canvas, Slide2d(ctx, resolveImage));
     }, this);
 
   },
