+++ lib/MediaLoader.js
@@ -1,9 +1,15 @@
 var assign = require("object-assign");
 var EventEmitter = require("events").EventEmitter;
-var hash = require("./loaders/hash");
-var loaders = require("./loaders");
+var hash = require("./hashResource");
 
-function MediaLoader (parallelThreshold) {
+/*
+A wrapper on top of loader, that:
+ - cache results
+ - maximize parallel loading
+ - provide events
+*/
+function MediaLoader (loader, parallelThreshold) {
+  this.loader = loader;
   this.resources = {};
   this._loading = {};
   this._threshold = parallelThreshold || 3;
@@ -19,6 +25,11 @@ MediaLoader.prototype = assign({}, EventEmitter.prototype, {
     }
     this._loading = {};
     this._queue = [];
+
+    this._imageResolver = function (url) {
+      var res = this.get(url);
+      return res && res.image || null;
+    }.bind(this);
   },
 
   // TODO Split the semantic of "hasPending" and "hasReady". same for StoreTransitions
@@ -37,11 +48,15 @@ MediaLoader.prototype = assign({}, EventEmitter.prototype, {
       this._load(descr);
   },
 
+  getImageResolver: function () {
+    return this._imageResolver;
+  },
+
   _load: function (descr) {
     var key = hash(descr);
     if (key in this._loading) return; // Already pending
     var self = this;
-    this._take(descr, loaders.video.call(this, descr, function (res) {
+    this._take(descr, this.loader(descr, function (res) {
       if (self._release(descr)) {
         self.resources[key] = res;
         self.emit("load", descr, res);
