+++ lib/SegmentKenBurns.js
@@ -1,12 +1,11 @@
 var assign = require("object-assign");
 var SegmentTimeline = require("./SegmentTimeline");
 
-var loaders = require("./loaders");
 var rectClamp = require("rect-clamp");
 var rectCrop = require("rect-crop");
 var rectMix = require("rect-mix");
 var BezierEasing = require("bezier-easing");
-var hash = require("./loaders/hash");
+var hash = require("./hashResource");
 
 var firstInstanceKey = "__skb_fi";
 
@@ -65,42 +64,30 @@ SegmentKenBurns.prototype = assign({}, SegmentTimeline.prototype, {
     };
   },
 
+  _setRes: function (result) {
+    if (result.video) {
+      this.video = result.video;
+      this._computeVideoMeta();
+    }
+    else if (result.image) {
+      this.image = result.image;
+    }
+  },
+
   _load: function (ctx) {
-    var self = this;
+    var media = ctx.media;
     var data = this.data;
-    if (this._loaded) return;
+    if (this._loadHandled) return;
     if (this.isFirstInstance) {
-      var res = ctx.media.get(data.image || data.video);
+      var res = media.get(data.image || data.video);
       if (res) {
-        if (data.image) this.image = res;
-        else if (data.video) {
-          if (res.video) {
-            this.video = res.video;
-            this._computeVideoMeta();
-          }
-          else if (res.image) this.image = res.image;
-        }
-        this._loaded = true;
+        this._loadHandled = true;
+        this._setRes(res);
       }
       return;
     }
-    this._loaded = true;
-    if (data.video) {
-      this._res = loaders.video(data.video, function (result) {
-        if (result.video) {
-          self.video = result.video;
-          self._computeVideoMeta();
-        }
-        else if (result.image) {
-          self.image = result.image;
-        }
-      });
-    }
-    else if (data.image) {
-      this._res = loaders.image(data.image, function (img) {
-        self.image = img;
-      });
-    }
+    this._loadHandled = true;
+    this._res = media.loader(data.video || data.image, this._setRes.bind(this));
   },
 
   ready: function (ctx) {
