+++ lib/hashResource.js
