+++ lib/loader/fallbackBlack.js
@@ -2,10 +2,10 @@
 var image = new Image();
 image.src = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNg+A8AAQIBANEay48AAAAASUVORK5CYII=";
 
-module.exports = function (loadingFunction, wrapped) {
+module.exports = function (loadingFunction) {
   return function (arg, success) {
     return loadingFunction(arg, success, function () {
-      success(wrapped ? { image: image } : image);
+      success({ image: image });
     });
   };
 };
