+++ lib/loader/index.js
@@ -1,0 +1,9 @@
+var fallbackBlack = require("./fallbackBlack");
+var timeout = require("./timeout");
+var loader = require("./loader");
+
+module.exports = {
+  loader: loader,
+  fallbackBlack: fallbackBlack,
+  timeout: timeout
+};
