--- lib/loaders/index.js
@@ -1,8 +1,0 @@
-var blackImageFallback = require("./blackImageFallback");
-var image = require("./image");
-var video = require("./video");
-
-module.exports = {
-  image: blackImageFallback(image),
-  video: blackImageFallback(video, true)
-};
