+++ package.json
@@ -41,7 +41,7 @@
     "rect-clamp": "0.0.0",
     "rect-crop": "0.0.0",
     "rect-mix": "0.0.0",
-    "slide2d": "^1.0.12",
+    "slide2d": "^1.1.0",
     "webglew": "^1.0.5"
   },
   "devDependencies": {
